This is a followup to #3810 which:
- removes the bidirectional matching code and `IsNotMatch` (still todo!) as the substitution produced by this code will always be a matching one
- re-factor he list and map matching code adding comments describing the match cases in more detail