@goodlyrottenapple 2024-04-18
> List matching looks good. One could argue that we should return larger chunks of the unmatched lists when we fail (to make it easier to identify the failed match in the given configuration). The `Map` matching is rather hard to understand IMHO.

I was thinking the exact opposite. If we have a bug/missing case in map/list matching, this returns the exact point of failure, making it (hopefully) easier to diagnose