@jberthold 2024-04-18
> > List matching looks good. One could argue that we should return larger chunks of the unmatched lists when we fail (to make it easier to identify the failed match in the given configuration). The `Map` matching is rather hard to understand IMHO.
> 
> I was thinking the exact opposite. If we have a bug/missing case in map/list matching, this returns the exact point of failure, making it (hopefully) easier to diagnose

OK, that's of course true: if we return a larger chunk the reader of the error message needs to follow the algorithm to understand where it hit the actual problem. I just noticed that in many cases the error will involve an empty list on one side... We can leave the error messages as they are now and see how this turns out in practice