@ehildenb 2024-04-19
Whatever we do, make sure the list and match matching algorithms are as _isolated_ as possible. Eventually we'll want ot reimplement them with standard algorithms based on Diophantine solvers from literature.