@goodlyrottenapple 2024-04-25
KEVM

| Test                                                            | sam-match-cleanup time | master-c391d6b87 time | (sam-match-cleanup/master-c391d6b87) time 
|-----------------------------------------------------------------|------------------------|-----------------------|-------------------------------------------
| benchmarks/storagevar02-overflow-spec.k                         | 40.79                  | 49.2                  | 0.8290650406504064                        
| benchmarks/storagevar02-nooverflow-spec.k                       | 41.6                   | 49.95                 | 0.8328328328328328                        
| benchmarks/ecrecover00-siginvalid-spec.k                        | 49.28                  | 57.85                 | 0.8518582541054451                        
| erc20/ds/transfer-failure-2-b-spec.k                            | 47.71                  | 55.86                 | 0.8540995345506623                        
| benchmarks/storagevar01-spec.k                                  | 40.88                  | 47.06                 | 0.8686782830429239                        
| benchmarks/ecrecover00-sigvalid-spec.k                          | 50.59                  | 58.15                 | 0.8699914015477215                        
| kontrol/test-expectreverttest-test_expectrevert_bytes4-0-spec.k | 87.82                  | 99.12                 | 0.8859967715899918                        
| mcd/dsvalue-read-pass-summarize-spec.k                          | 52.2                   | 58.05                 | 0.8992248062015504                        
| benchmarks/storagevar03-spec.k                                  | 40.54                  | 45.02                 | 0.900488671701466                         
| erc20/ds/transfer-success-2-spec.k                              | 62.2                   | 68.76                 | 0.9045956951716114                        
| erc20/hkg/transferFrom-success-2-spec.k                         | 55.81                  | 61.21                 | 0.9117791210586506                        
| mcd/flipper-ttl-pass-spec.k                                     | 55.83                  | 61.19                 | 0.9124039875796699                        
| erc20/hkg/transferFrom-success-1-spec.k                         | 60.55                  | 66.18                 | 0.9149289815654275                        
| erc20/ds/transferFrom-failure-1-c-spec.k                        | 98.46                  | 106.09                | 0.9280799321330945                        
| mcd/vat-slip-pass-rough-spec.k                                  | 93.92                  | 100.82                | 0.9315611981749654                        
| mcd/vat-move-diff-rough-spec.k                                  | 168.8                  | 180.8                 | 0.9336283185840708                        
| benchmarks/storagevar00-spec.k                                  | 41.07                  | 43.32                 | 0.9480609418282548                        
| functional/evm-int-simplifications-spec.k                       | 113.4                  | 119.5                 | 0.9489539748953976                        
| examples/erc20-spec.md                                          | 229.05                 | 220.09                | 1.0407106183833887                        
| mcd/flipper-addu48u48-fail-rough-spec.k                         | 54.96                  | 52.64                 | 1.0440729483282676                        
| mcd/cat-exhaustiveness-spec.k                                   | 109.94                 | 104.38                | 1.053266909369611                         
| mcd/vat-subui-pass-rough-spec.k                                 | 126.51                 | 119.39                | 1.0596364854677947                        
| erc20/ds/transfer-failure-1-a-spec.k                            | 75.36                  | 70.42                 | 1.070150525418915                         
| erc20/hkg/transfer-failure-2-spec.k                             | 60.7                   | 55.39                 | 1.0958656797255824                        
| erc20/ds/transferFrom-failure-1-a-spec.k                        | 90.88                  | 82.14                 | 1.1064037009982954                        
| erc20/ds/transferFrom-failure-2-c-spec.k                        | 60.42                  | 54.52                 | 1.1082171680117388                        
| mcd/flipper-tau-pass-spec.k                                     | 62.6                   | 56.02                 | 1.11745805069618                          
| erc20/hkg/balanceOf-spec.k                                      | 47.68                  | 42.15                 | 1.1311981020166073                        
| erc20/ds/transferFrom-failure-1-d-spec.k                        | 70.66                  | 61.69                 | 1.1454044415626519                        
| kontrol/test-expectcalltest-testexpectstaticcall-0-spec.k       | 58.85                  | 48.88                 | 1.2039689034369885                        
| TOTAL                                                           | 2249.0599999999995     | 2295.840000000001     | 0.9796240156108433